Preperation for Version 0.1.0 - major makeover
*new file structure

The plugin has become bigger as expected during the first year. To stay organized, the previous idea of a singlefile is dropped and the code is separated by topics.

*performance

The previous versions have some issues with complex structures. Instead of single curves, one connected mesh is created for visualization. This is way faster and also memory efficient.

Scipy is used if available. This feature is available for experienced users that are able to install Scipy into Blender. A checkbox use_scipy is integrated, if the user wants to disable Scipy in case of errors.
saving data

Previous approach was to store all data into python classes. This data has been lost, if the file was closed. The new version is going to use a new variable system within bpy.context.scene[“<Phaenotyp>”]. All data is stored within the blend-file and is used when reopening a saved file.

The way how data is stored in variables has improved in a lot of parts.

Color and Radius of the visualization-mesh is stored in vertex-groups and vertex-color-attributes. This is going to be quicker and also more organized.

*processing data

Phaenotyp is switching to the amazing geometry-nodes in a lot of tasks. A lot of stuff is way more organized and quicker.

*genetic algorithm

The genetic algorithm has been reworked in a lot of parts to work more efficiently.
The genetic algorithm can perform a sectional optimization before running the fitness-function for each individual. In this way the overall exploitation of the structure can be increased.
visualizing data
After using the genetic algorithm the result can be ranked by best fitness.

Due the new approach of processing data, the visualization ist way quicker and more memory efficient.

*usability

Some parts are more user friendly hopefully. All experimental functions that were somehow confusing have been removed.

Defined parameters are grayed out after a successful calculation of the structure. In this way, the user can not change parameters, that would not affect the finished result. Otherwise this could lead to some confusion, if the structure is saved with wrong parameters.

*variables

Because of the improved way of storing data, the variables have been reworked in a lot of parts. The overall management of variables is redone.

*hints

In order to install this new version, any older one needs to be uninstalled. A description for users is going to be added.

There is a division by zero error in sectional optimization sometimes that needs to be fixed or excluded with try/except.