Two types of mating are now available
The existing type is now called "direct". The chromosome is generated by taking the genes of the parents directly. It looks like this method creates a high diversity of generated individuals. The overall evolution tends to jump between tendencies.

The added mating type is morphing the genes of parents. It looks like this is generating less diverse populations and tends to stick to one tendency. The fitness of a solution can be higher within this tendency.