Update __init__.py
Warning is added if the user is choosing a value for elitism that is higher than 50% of the population_size. This would be useless because the new generations would always be the same. Furthermore it would lead to an error in the generation, because the key is not in subgroup dict.