minimal value of Do and Di
When sectional performance is running for multiple times, the diameter of a member can have a value of zero. This will lead to a division by zero.

In order to keep the ratio of Di and Do a temporary variable called Do_Di_ratio is used.

If the Di of a member is below 0.001, the Di is set to 0.001. In order to avoid another division by zero in the calculation of