changed some routines to improve usability and flow
next to [str(frame)] a key called ["first"] is created to keep track of the input from the user. This is importand to avoid errors if the user is chaning the frame during the setup of the material for some reason.

update_post created an error if Phaenotyp is not used at all. Also an error was dropped when there has not been a member or a result at this specific frame.