bruteforce
New function called bruteforce is integrated. Idea is to have a possiblity in which the ga is just pairing all possible solutions. This can be helpfull if there are only two or three shape-keys. Furthermore it is usefull for debugging and finding the overall best possible solution to compare to the ga.

It is integrated as mating-type. From a technical point of view it is working in a different way but from a logical point of view it is somekind of mating if all members a are paired in all possible combinations.

The matrix of possible combinations is created with ittertools.