report is showing Do, Di and A
The first entry of Do, Di and others was stored as key "first" with all other entries. It would have been complicated to integrate this in the report, because the dict is sorted by the frame id. There is a new entry Do_frist ... to avoid this issue. This looks also way cleaner for other operations.