added new id prop called frames
id prop frames is used to store all info about the structure at a specific frame. The area, length, kg, volume and other infos are stored in this dict. This is also for fintess functions. The reason why this infos are separated in a new dict is, that this infos can only be gathered with frame change. In order to iterate the frame change once in prepare_fea.

Other parameters regarding a specific frame can be stored here in future.