buckling resolution
Is used to define the amount of members that formaly build a curve. This is needed if a user is convertig a curve to a mesh with a resoltion other than 1. In this case, the curve is splitted up in seperate segements. In order to calculate the buckling correctly, the resolution needs to be defined in advance.